worker/requirements.txt
fastapi
pymongo
mysql-connector-python
python-dotenv
pandas
numpy
celery
bcrypt
requests
